More code quality improvements
 - add exception-tracing to `LoggingHandler`
 - more use of `std::size_t` and fixed width integer types,
  `std::int64_t` etc.
 - use `std::array<Vector3, 3>` for `bravais_vectors` in `Geometry`
 - improve `LoggingHandler` performance by turning some function
  parameters into `const &`s
 - fix some comments

+++ core/src/Spirit/IO.cpp
@@ -230,7 +230,7 @@ try
                 "OVF file \"{}\": segment {}/{} contains only {} spins while the system contains {}.",
                 filename, idx_image_infile+1, file.n_segments, segment.N, image->nos),
                 idx_image_inchain, idx_chain );
-            segment.N = std::min(segment.N, image->nos);
+            segment.N = std::min(static_cast<std::size_t>(segment.N), image->nos);
         }
         else if( segment.N > image->nos )
         {
@@ -239,7 +239,7 @@ try
                 "Reading only part of the segment data.",
                 filename, idx_image_infile+1, file.n_segments, segment.N, image->nos),
                 idx_image_inchain, idx_chain );
-            segment.N = std::min(segment.N, image->nos);
+            segment.N = std::min(static_cast<std::size_t>(segment.N), image->nos);
         }
 
         if( segment.valuedim != 3 )
@@ -525,7 +525,7 @@ try
                         "OVF file \"{}\": segment {}/{} contains only {} spins while the system contains {}.",
                         filename, start_image_infile+1, file.n_segments, segment.N, image->nos),
                         i, idx_chain );
-                    segment.N = std::min(segment.N, image->nos);
+                    segment.N = std::min(static_cast<std::size_t>(segment.N), image->nos);
                 }
                 else if( segment.N > image->nos )
                 {
@@ -534,7 +534,7 @@ try
                         "Reading only part of the segment data.",
                         filename, start_image_infile+1, file.n_segments, segment.N, image->nos),
                         i, idx_chain );
-                    segment.N = std::min(segment.N, image->nos);
+                    segment.N = std::min(static_cast<std::size_t>(segment.N), image->nos);
                 }
 
                 if( segment.valuedim != 3 )
@@ -1076,7 +1076,7 @@ try
                     "OVF file \"{}\": segment {}/{} contains only {} vectors while the system contains {}.",
                     filename, idx+1, file.n_segments, segment.N, image->nos),
                     idx_image_inchain, idx_chain );
-                segment.N = std::min(segment.N, image->nos);
+                segment.N = std::min(static_cast<std::size_t>(segment.N), image->nos);
             }
             else if( segment.N > image->nos )
             {
@@ -1085,7 +1085,7 @@ try
                     "Reading only part of the segment data.",
                     filename, idx+1, file.n_segments, segment.N, image->nos),
                     idx_image_inchain, idx_chain );
-                segment.N = std::min(segment.N, image->nos);
+                segment.N = std::min(static_cast<std::size_t>(segment.N), image->nos);
             }
 
             if( segment.valuedim != 3 )
